refactor(desktop): convert main process to Effect-TS
